Transpose Board Cells in the view for mobile browsers
This still needs some work... but it's a start. Basically what we're
trying to do here is:
1. Be able to detect if the current browser is a Mobile browser,
2. Feed that ability into our Grid object, which is what's responsible
   for transforming our list of Cells (one long Array) into an Array of
   Arrays (one Array per row).
3. Transpose the Array of Arrays if it is a Mobile browser.
   Transpose as in: https://en.wikipedia.org/wiki/Transpose
   - Which Ruby makes very easy with the Array#transpose method.

I foresee other options playing into the mix later, like... only
transpose if the Game is larger than the Beginner board size, then
transpose the board. Otherwise leave it alone. The reason being that
we're only transposing to make it easier to view large Game boards on
mobile (in portrait mode vs in landscape mode). It actually doesn't end
up mattering a lot if we just always transpose for Mobile... but I don't
love the thought of it.

After transposing the board, it still doesn't fit on the mobile browser
size... so we have to add horizontal scroll bars for the Game board.
That is, instead of having the board add a bunch of margin on the right
side of the whole page and using the horizontal scroll bar for the whole
page, we can limit it only to the Game board itself. This allows the
text above and below the Game board to stay put while scrolling and
feels much better than a broken looking page.
- To support this, I had to isolate the table header and footer out into
  their own divs.
  - To support this, I had to perform a miracle. (It's hard to make
    page elements size themselves according to the sizes of other
    elements.)
    - The miracle comes from the container element:
     `<div class="w-fit max-w-full">`. This tries to be as wide as it
     needs to be `w-fit`, but only to a point `max-w-full` (that point
     being at the 100% of the page width). I'm actually surprised it
     worked. ¯\_(ツ)_/¯

I'm not yet happy with:
1. The view template passing the context into the View Model's `rows`
   method...
2. The context that's passed in being so large in scope. I'd like to
   limit this to just what's needed, via a simple PORO.
3. There's not yet a way to test this in non-mobile devices. I'm
   thinking about adding a Feature Flag service to allow conditionally
   enabling this anywhere, for testing.
   - Or we could just make this a legit feature of the game? i.e. a
     "transpose" button in the UI?