Make Grid into an Enumerable object
This just gives us more functionality for free. e.g. As an Enumerable,
we automatically get #each, #to_a, #first, #count, #any?, and so on.
See: https://ruby-doc.org/3.3.4/Enumerable.html
- Note: We are missing out on a few of the Array-like methods... which
  I'm OK with for now. Those include e.g. #size, #length, #[], #empty?,
  #last. We can add them if we need them later. We would just forward
  these messages on to our #to_a method (which comes from Enumerable
  now).

We can take advantage of this Enumerable-ness in the Games::Show View
Model. We no longer need to call `board.grid.to_a` in order for the
view template to be able to iterate over the grid rows for rendering.
Now, just `board.grid` will suffice because it's already enumerable!