Fix: Don't attempt a Cell reveal on mouse drag
Previously, I was noticing occasional 404 errors when "clicking?" on the
board. I wasn't sure. But finally nailed it down to a click/drag. This
seems to set the `event.target` as the `<tr>` (row) element instead of
the `<td>` (Cell) element. So, we can't just "nope" out of the reveal
event handler if we're not dealing with a Cell.