Don't broadcast refresh on DutyRoster dec/inc events
This is a bit unfortunate, because it added a sense of "live"-ness to
the Players Count indicators... but I noticed that broadcasting refresh
on the `current_game` channel results in updating even just the "New
Mission" view. i.e.
- if I'm on the Home page, which is currently showing the "Do you have
  what it takes?" + Game start buttons, and
- someone else is too and just refreshes the page
- or if someone else just visits the site anew at this time
... this causes a disconnect/reconnect event to the Turbo Stream, which
decrements/increments the Duty Roster, which broadcasts refresh to the
`current_game` channel which refreshes everyone's browser.

I considered using a different Turbo Stream channel name outside of
`current_game`, but... this being the same name is what makes it so
that all visitors see the same flow in the "War Room":
- When a game ends everyone sees it.
- When someone starts a new Game, everyone sees it.

One thing we did lose from this (which was a byproduct of the "bug" that
caused me to remove it), is the ability to see when someone goes to the
"New Mission" page. i.e. that broadcast doesn't happen anymore. I'm not
sure that's good or bad yet. But we could add that feature back to the
controller actions somehow later.