Add tests for App module
Testing the backwards logic was too hard, so I reversed it. i.e.
`include_test_difficulty_level?` is easier to reason about than.
`exclude_test_difficulty_level?`

I also realized, through testing, that the previous attempt at
memoization wasn't going to work since `a ||= false` will run every
time. So we have to do an explicit Nil check, and only run the logic
if the class variable is `#nil?`.

Next, I started seeing random testing errors, which is never good or
fun to track down. But I had a suspicion straight away that it was due
to how we're memoizing state into the App class during the App test
examples. So, i added a `before` block to pre-set that class variable
back to `nil` before each test example run.

However, that wasn't all! There was another place where class state
was being affected. Which was in the DifficultyLevel::SETTINGS_MAP
constant. Since we conditionally exclude the "Test" Difficulty Level,
based on the result of our method under test:
App.include_test_difficulty_level?. The easiest way to fix this is to
lazy-initialize the "Settings Map" in Difficulty Level. i.e. don't load
it as class load time, but rather, only after it is first accessed.
This allows the "running" class to determine whether or not  to include
the "Test" difficulty level vs the "loaded" class. Annndddd... this is
part of hwy lazy loading is one of the tenants of OOD. i.e. putting off
loading a thing until the last second.
- We didn't run into this before because before, our memoization of
  App.include_test_difficulty_level? wasn't working properly. After
  fixing it, we started seeing Class-level state issues in
  DifficultyLevel.