Don't require a "Seed Cell" when placing mines on Board
It's perfectly legit to default the "Seed Cell" to `nil` when placing
mines on the Board. It's less legit to do so when starting the Game,
though, I think. i.e. want to force the caller of Game#start to think
about sending a Seed Cell in, even if they do choose `nil`.

This amounts to just a slight simplification on the Board#place_mines
tests... but also helps out when playing around in the Console a bit.