Add a basic UI Portal
The new UI Portal is available, only in the `development` env, at: `/ui`

Here, we can display all of the Flash types for easily testing them.

Also, a list of all Icons in the system, etc., so that we can quickly
know what is available to us, and what they look like.

And more to come...