Use requestjs-rails for Cell actions POST submits!
I stumbled on requestjs-rails today. Which is a library that
encapsulates the logic to send Rails headers, like the CSRF token,
Turbo Frame ID, etc, in JS-initiated (`fetch`) requests. See:
https://github.com/rails/request.js
- We can use this to totally replace our `boardActionProxy` form and
  associated data massaging and submission via board_controller.js.
  I can't believe the existence of this isn't better documented... it's
  exactly what I was searching for!

Adding this JS library involved a gem install, so I updated the rest of
the gems at the same time (as dependabot was wanting as well).

References:
- https://www.betterstimulus.com/turbo/form-submits (where I found out
  about requestjs-rails).