Use `<time>` element for Game Timer
I'd forgotten about the existence of the HTML `<time>` element. We can
use this for better semantic clarity in our HTML. Plus it gives us an
opportunity to use its associated `datetime` attribute for even more
context.

For the `datetime` attribute, we use the format string:
`"PT<elapsed_time_in_seconds>S"` which means:
- P = the start of a duration (Period)
- T = indicates that a Time part is following (vs a date part)
- S = "seconds"

We could be a bit less cryptic in our format by separating the seconds
out into Hours/Minutes/Seconds, but I'm not sure the juice is worth the
squeeze on writing that code. Can do so later if there's good reason to.

I've also taken the time to include an `aria-label` for the `<time>`
element to add yet further meaning to the element. By adding an
"Elapsed time: <01:23:45>" label. This is an easy-to-provide
accessibility feature so 👍.

References:
- https://www.html-easy.com/references/html-time-tag/