Add CellTransactions Models / DB tables
Add the basic structure for Cell Transactions, including:
- Cell Reveal Transactions
- Cell Flag Transactions
- Cell Unflag Transactions

Cell Transactions are associated with Users and Cells and have the usual
created_at/updated_at timestamps.

These aren't hooked up to anything yet--so they won't yet be created--
but the associations are present at this point for querying.

Note: Care is taken that when a Game is deleted, the associated
CellTransaction records are too. But not so for User deletion. We prefer
to keep CellTransaction history even if a User is deleted because they
record the history of events on the Game still, even if we don't know
who did them. i.e. we can just call it "Unknown User" doing them.