Create Cell*Transaction records when appropriate
When the Cell Actions controllers run and/or when their associated
Service Objects run, we can now go ahead and create the Cell*Transaction
records to record these events.

To support this, we also need to pass into the Service Objects the
"current User". Instead of adding a 4th argument to initialize the
Service Objects with, I've chosen to pass in a new CurrentContext
object. This is just a tiny PORO that plays the part of knowing about
all of the context-related objects needed for the Service Objects to
play their roles. I could have just passed in the controller context
instead--and I almost did--but this is too much. Using the new
CurrentContext object instead limits the scope.

Also:
- I wanted to organize Transactions separately from `models`. To
  accomplish this, we have to tell Rails to eager_load the
  models/transactions/* path manually. Why? Because we're going outside
  the norm, which is that in order for Rails to look into a subfolder,
  there must be a module designating the need for it. e.g. if our
  CellTransaction objects were called Transactions::CellTransaction,
  etc., then we wouldn't need to add this path to the eager_load list.