Designate CellTransaction as an Abstract Base Class
We don't ever want to be able to instantiate a new CellTransaction
object. It doesn't make any sense to. There's no way to then save one,
for example, because we require a value for the `type` column. So, we
can mark it as an Abstract Base Class to prevent e.g.
`CellTransaction.new` from working in the first place.

We use a custom `as_abstract_class` macro to accomplish this--define in
the new mix-in: AbstractBaseClassBehaviors. We have to do this vs e.g.
using Rails' `primary_abstract_class` macro, which does too much. Rails'
version actually breaks our subclasses in strange ways if we try to use
it.