Add ability to update username on Game end
Add the routes, controller, model validations, CSS rules, view updates,
locale entries, etc to get User#username editing working. For now, at
least (maybe for always?) the opportunity to edit ones username comes
only at the end of the Game, and only for Games that they've
participated in. In our case, by showing up in the Duty Roster, we
know that the User participated in the Game and so we just need to check
if the user is the `current_user`.

Note:
- Editing usernames isn't using Turbo yet, but this is step one: get it
  working normally. Then, getting it working with Turbo can be
  progressive enhancement.