Show New Game buttons on Game End
This is better/quicker/faster than the old "Restart" button that just
refreshes the home page. Now, we can start a new Game in just a single
click!

While working on this, I tested by lengthening the amount of time it
takes for a Game to be considered "Just ended" / "No longer current" up
to a day (from `1.second`--See Game::DEFAULT_JUST_ENDED_DURATION). When
doing so, it became apparent that we actually do want to take the `last`
game within Game.current instead of just `take`ing an un-ordered one.
This way, we deterministically get the most-recent (current) Game. Not
that this will ever happen, as is, but it's still more technically
correct this way.