Massively speed up new-Game/Board generation
This commit adds Board::Generate -- a Service Object that now inserts
all of the {Cell} records for the Board into the Database via
`Board.insert_all`. This is the Rails bulk-insert method that was added
in Rails 6. Doing this gives us a massive optimization over adding
each {Cell} record to the {Board#cells} collection one-by-one and then
saving them to the Database one-by-one like we used to do.

Building an Expert-sized Board:
Before: 2.5 seconds (ActiveRecord: 296.9ms (484 queries) | GC: 79.1ms
After 0.08 seconds (ActiveRecord: 29.0ms(5 queries) | GC: 6.6ms)

So basically it feels instant now. 👍

The biggest difference with this approach is that we can't "build" Cells
for a non-persisted Board. So first the Game and Board must be created,
and then we can bulk insert the Cells based on the Board#id. But we can
at least use a DB transaction to wrap the creates to ensure that if
either of them works they both do.