Add ActiveRecord validations + tests to User & Cell
We were just previously lacking validations and/or tests.

To support this effort, I've added a test support class:
ValidationError. This aids in looking up I18n error messages by name
in validation-related tests. This module is imported and heavily adapted
from an old project.

After adding validations on Cell#value, we finally run up against our
old decision to inject view-level understanding into Cell#value model
method. There, we attempted to output `🚩` if Cell#flagged? was true...
but, again, this is a view-level concern. `🚩` is not a valid
Cell#value. (Come to think of it... maybe `💣` shouldn't be a valid
Cell#value either...)
- This issue came up with some failing tests in CellTest. It wasn't
  clear what was wrong for a while. To fix this up for the future, I've
  updated all of the <Model>.update() calls within tests to use
  `update!` instead--which will fail fast by raising an exception if
  there is a validation issue.

Also:
- Add a uniqueness constraint on cells.coordinates (x/y pair) to the
  database. This seems more critical now that we're doing bulk insert
  of Cell data.
- Adding User#username validation tests exposed a potential bug in the
  User#username= setter, when given `nil`. So, fixed this as well.