Refactor CoordinatesType -> Type::Coordinates
For me, this is actually clearer as it follows the ActiveModel::Type::*
style name spacing, upon which we're building. It's also in preparation
for a new Type, coming in a following commit. So now we'll have more
than one custom attributes Type... and so an organizational folder for
them all seems fitting as well.