Refactor Board#[columns|rows|mines] -> Board#settings
In general, instead of passing around a slew of related attributes, it's
better to have an Object that represents them. So this commit takes the
opportunity to combine the independent Board columns/rows/mines
attributes into a Board::Settings object. At the same time, we change
the concept of columns/rows to width/height as the latter is much more
natural to think about.

Note:
- The implementation for Board::Settings and Type::BoardSettings is
  very similar to Coordinates and Type::Coordinates. But, I'm resisting
  any urge to abstract the similarities between these concepts at this
  point. Not enough information still... and it may never even be
  prudent.
- Board::Settings is pretty similar to DifficultyLevel...but they're
  different concepts. That said, I do still want to refactor
  DifficultyLevel away from using columns/rows and towards width/height.
  I *still* can't quickly think about columns/rows as easily as
  width/height. So... refactor for this is coming later.