Add CSS external link indicator + helper
We can automatically have anchor tags whose href starts with `http` be
styled as external links. This is great for e.g. links in en.yml.

On top of this, it's also useful to have `rel="noreferrer noopener"` in
external links. There are potential security issues with opening links
in a new tab where the opened site can refer back to the source
(referer/opener) and do malicious things.

Browsers today are supposed to automatically disable this behavior for
links with `target="_blank"`, but I'm not sure I fully trust this, so
I want to include the `rel` attribute anyway for all external links.
- In support of the above, we add the `external_link_to` helper method.

Notes:
- Unfortunately, I can't find a way to style the external link opener
  SVG icon as a background image. So, I punted and just made 2 different
  version of the SVG. The normal one, and one colored gray for use in
  the Dark theme. These are added to our Tailwind config for ease.
- Also: Break `site_description` and `site_description_html` into
  `site.description` and `site.description_html` in en.yml. Nesting
  feels more natural, and invites other keys in the future.

References:
- https://techstacker.com/html-noopener-attribute/
- https://css-irl.info/styling-external-links-with-attribute-selectors/