Make Site Description banner be permanently dismissable
We will now store a cookie to record if/when a user has dismissed the
Site Description banner at the top of the home page. This text is later
findable in the "About" page. It is still good for it to appear on the
home page as well, though, since it provides an immediate "about"-lite
description of what the site even is.

Notes:
- The cookies.js library was mostly written by Grok. It works! I put it
  in a `lib` folder because I'm not sure what better to do yet.
- The only way to get the library back is to delete the cookie. This
  can be done in the developer tools section of the browser.