Fix up overall margins/padding of main-page elements
Previously, there was an odd space between the header/footer borders
and the actual side of the browser window. This commit fixes that by
removing the `container` from the header/main/footer elements and,
instead, wrapping all of the pages' contents in `container` divs. This
way the header/main/footer can actually reach the edges of the browser
window but the page content can't.

Also, squeeze in a little bit more space for the mobile Game Board view
by adding negative-x margin to the game board / table.