Add `rake eslint` task + fix found JS issues
Also, fix config for eslint so false positive errors aren't reported in
tailwind.config.js. Most important was telling eslint to run in `node`
mode, so that the `require` statements aren't flagged as errors. But,
no fix I could see for the `defaultTheme` not used issues, so just
added it as an ignore.

Also, add `rake eslint` to the default `rake` task list.