Inline the username edit feature, using Turbo
Had to switch from an `<ol>` to a manual list with `<div>`s to preserve
intended formatting. Having a `flex` element inside of an `<li>` seems
to break it. It turns the `li` into a multi-line thing with the number
being on the first line and then the contents being on the 2nd line.

Also, to simplify logic and rendering, we abandon `link_to_if` for this
username-edit link.