Add User - Show links to all users in Duty Roster
To support this, I've made just the pencil be the edit / "sign" link
for the current user.

Also, use an old hack to make the username field auto-select text on
form show. Better would be to use a Stimulus controller and do the same
on `connect`, but this is so... easy. Going to try it out for a bit.