Prevent dismissing welcome banner until User signs name
We'll just no longer show the "X" (dismissal/close button) on the
Site Info / welcome banner on the home page until the current User has
signed their name to a game. This services as both a bit of a "thank
you" to the User, as well as prevents premature dismissal (before the
User actually got a chance to read it... if they're like me and just
"X" out of alerts without reading all the time.)

To support this, I was doing down the route of adding anew flag to the
ApplicationBanner... but it got too confusing very quickly. Because
then we had e.g. `show_dismissable_button?` and
`permanently_dismissable?`. So the easy question to ask was: what's the
difference? Instead of typing out a huge comment, I decided to remove
the ambiguity by making a new class. This is the principle of: If you
have enough info to know how to switch on a behavior, then you have
enough info to just pass in a smarter thing. Or, in our case, a
specialized type: Application::PermanenetlyDismissablebanner (vs the
existing Application::Banner). Then, each of these objects can
effortlessly service the view's question: `permanently_dismissable?`--
just based on their type.

I've also taken this chance to simplify Application::Banner a bit. I'd
consider DRYing things up the next time I feel the need for a new type,
but for now I"m letting it slide even though it feels obvious how to
DRY these 2 Banner types up.