Move User name signing up above Game Stats on Game end
I've been struggling for the best place to insert username
signing/update on Game end. I think this is the best spot yet. It's
more front-and-center and harder to miss among all the other (now)
clickable usernames in the Duty Roster. Plus it offers a chance to
actually prompt for "signing" the user's name. Which is the feeling I've
been going for.

Also:
- Be sure to update all spots the username may appear on sign/update of
  username. No IU component left behind!
- Update the site welcome banner on user sign. This way, the
  close/dismissal button has a chance to appear, signifying that this is
  a direct result of (or unlock achieved by) signing.