Disable dependabot
This app is not under any threat from anyone and stores no personally
identifiable information. Plus the dependabot alerts seem to be much
more frequent than they used to be.... so not really worth it.

As an homage to dependabot--who does mean well!--I'm also updating the
gems at this time :). It wanted us to update puma.