Only transpose wide Grids (in mobile view)
Previously we would just always transpose the board on mobile, even if
started out as square or as wider than it was tall. This is a needless
transformation for square boards, and should be avoided. And, for tall
boards, this is actually harmful to the experience.

To accomplish this, we both the view context and some info about the
board/Grid shape. So the code has to shuffle around ab it, but I think
the result is actually much simpler. The only problem is we no longer
allow injection of a custom organizer. Doesn't matter for game play, but
as an OO purist, this hurts a little.

Note: We're also careful to only transpose the grid if it's wider than
10 columns--even on Mobile, since 11 columns still fits just fine on
mobile displays. (11 columns fits for me, but I'm being a little
conservative and choosing 10 as the limit instead.)