Only show "Sign" link if current user participated in Game
This was an oversight before. This protection was just lost in the
shuffle of moving the sign link out of the Duty Roster section. So now,
we need to have an additional check to only present the "Sign your
name?" link/option if the current_user actually did participate in the
Game.