Fix banner-close cookie storage in Safari
Firefox is too lax about storing a secure cookie on a non-secure
(non-ssl) domain/request. Safari is strict in this. So, cookies weren't
being stored in Safari (on development at least).

So this commit changes cookie storage to automatically check for whether
or not "Secure" can be set, and does so if possible.

Note: I've also changed the "Permanent" cookie duration to 1 year, down
from 20. This is also based on Safari, which seems to allow a maximum
of 13 months. 1 year is fine.