Don't render Game board within .container div
Rendering the Game boar inside of the standard `.container` div makes it
artificially ugly on narrow screen widths--it shows a left-margin while
the right side of the board is overflowing off the screen. And removing
the game board from the `.container` element is fine because we already
use `mx-auto` to center it.