Fix partial theme change on turbo "refresh" in mobile
Before this change, mobile clients that are using dark mode would see
the top-bar of the browser / status area change from the dark theme back
to white on turbo_stream.refresh events. i.e. every time a Cell is
clicked on from another session.

We fix this buy adding our (what is now called) setTheme() function on
all `turbo:load` events. This is the best solution I can find, anyway.

Note: I've tried moving all of this JS out to a pinned module, etc, but
"module"-style scripts (not sure how to best use this terminology) are
deferred from loading until the DOM has loaded. This is too late and
creates a FOUC (Flash of Unstyled Content). So we're still stuck with
a `<script>` tag sitting in the `<head>` of the application layout.