Massively speed up recursive reveal of blank cells / pops
Previously, we used straight database queries for getting neighboring
Cells, queried by constructing a set of neighboring Coordinates. But,
after noticing some pretty long load times for recursive reveals I got
to thinking there were just too many ActiveRecord queries behind this.

This commit removes most queries associate with:
- Getting neighboring Cells
- Recursively revealing neighboring Cells (via bulk upsert)

Neighboring Cells are now selected just with Ruby `select`. And so on
for the supporting methods -- all straight Ruby now. This ends up being
much (much) faster. Largely because we already have the Board#cells
loaded... it's just a big Array. Anyway... here's the results:

GIVEN: A 20 x 20 grid with 1 mine on it. Where revealing 1 cell
recursively reveals all Cells and ends the Game in victory.
- Before this change:
   21.2 seconds
   (ActiveRecord: 6717.7ms (5801 queries, 0 cached) | GC: 499.7ms)
- After this change:
   0.4 seconds
   (ActiveRecord: 65.1ms (24 queries, 0 cached) | GC: 14.3ms)

So, we're now over 50 times faster!

While doing this, I also found that the RecursiveReveal code between
Reveal and RevealNeighbors could be the same if the lead-up in
RevealNeighbors was massaged a bit. So I pulled this service object out,
to be used by both controllers. This also fixed a bug in the
RevealNeighbors version of the Recursive reveal where any incorrectly
placed flags that were in the way would break the recursive reveal,
creating an awkward gap. Reveal wasn't doing this, and now neither of
them do.

WARNING:
- There are limit to recursion! Attempting to clear a 30x30 board that
  has just 1 mine is... too much. We get a StackLevelTooDeepError. I've
  tried to optimize the call stack, but we can only do so much. I plan
  to solve this by adding validations on mine density for Custom Game
  Board sizes.