Don't allow Custom Games with too high/low mine density
Also, combine validation efforts between Board and
CustomDifficultyLevel. These really are the same thing... just 2
different objects that need it. Luckily, ActiveModel::Errors.merge!
exists! -- for merging another object's Errors into self.errors. This is
perfects for what we need. CustomDifficultyLevel is better at formatting
of the error messages anyway, as it's the User-facing piece that serves
the form when creating a "Custom" Difficulty Level.