Remove the last of the Board#cells-related queries
Using straight Ruby-style enumerable methods ensures we won't get lost
in differing instances of the same Cell object. This is what was
happening in RevealsController, for example. Basically, we get one set
of Cells from one query and a different set from the main Board#cells
query. So mutating cells in one set doesn't affect the other, and it's
so hard to tell/reason about which you're dealing with. Best to get rid
of them all. Regardless, they've proven to be much more efficient.