Mark the Game Timer as data-turbo-temporary
I'm not sure I'm using this attribute properly, but just trying to see
if this helps with the occasional weirdness we'll see in the Game Timer
where it'll flop back and forth between a couple of time states.

At some point, the JS for this would best be rewritten to start from a
Date/Time and calculate difference from that instead of relying on the
1-second setTimeout interval to just increment the number of seconds.
I don't think this 1-second interval is reliable enough for a timer.
During the normal course of gameplay, it's fine since the server
refreshes the timer with each click, but it's inaccurate for lagging
games.