Fix incorrect display of Mines "count" on Game init
Previously, we were using a count of the current number of mines in the
Game Board for displaying the Mines "total". But, this should actually
be just a reference of the Difficulty Level setting. There's no need for
it to be actually tallied up. Tallying, for example, causes `0` to be
reported before the mines have been placed in the board (which happens
on the first Reveal).

So we get rid of Board#mines_count in general and just refer, instead,
to Board#mines (the attribute which comes from Board#settings). Then,
we can also simplify `any_mines?` to `mines_placed?`, where all we need
to know is if there are any mines present on the Board (yet).

Also, fix left/right margins on the Game Board header/footer. They
looked off because we've recently changed the board to use `-mx-5`, so
that it doesn't get any more hidden than necessary at browser widths
that are too skinny for the Board size. We can give the footer/header
a similar negative margin to get them back into proper position relative
to the Game Board.