Replace DifficultyLevel with Board::Settings
This is just a massive internal refactoring. There was some
DifficultyLevel-related functionality in Board, some in Game, to an
extent, etc. I think this Board::Settings implementation is a bit easier
and cleaner. It tries to do a little less. And at the same time is
perfectly accepting of "Custom" settings. Because... it's a Settings
object! It was painful needing a separate CustomDifficultyLevel before.
Plus the Conversions.DifficultyLevel conversion function was a bit much
for what we needed. I kept running into this but never really knew
what to do about it. I think this is better... but I don't know if I'll
ever truly be satisfied!

Still to come is refactoring Game#difficulty_level to something else...
And perhaps a refactoring on where CustomGamesController and
RandomGamesController live. etc.?