Allow Coordinates to be expressed as Ranges
To make Coordinates support Ranges, we just have to implement `succ` and
include Comparable. See:
https://ruby-doc.org/3.3.5/Range.html#class-Range-label-Ranges+and+Other+Classes

From this we get the ability to express Coordinates as a range along a
row. e.g.
```
Coordinates[0, 0]..Coordinates[30, 0] # => (0, 0), (1, 0), ... (30, 0)
```

While this has no practical implications for our App code, I think it's
a cleaner way to express rows of Coordinates in tests and on the
console.