Store/Get Custom Game settings to/from cookie
This is an effort to make it easier to start and restart custom Games.
By remembering the previous settings the User entered, they don't have
to (a) remember what they were last time, or (b) spend the time entering
them again and again.

Board Settings are stored in a cookie on successful submission, and then
retrieved the next time the form is loaded. We're careful not to force
these values on the form, though. e.g. if the user has edited them and
submitted the form and errors come back. We want the previously entered,
yet invalid, values to still be showing.

Also: Unify the mechanism for cookies storage. There's a little bit of
logic to this, so it's nice to abstract it.