Add (dev) Draw Mode + Export of drawn patterns
To enable Draw Mode:

```
DRAW_MODE=1 bin/rails s
```

Once in Draw Mode, clicking on the board will Flag a Cell instead of
revealing it. All the other click types are also disabled. The board
header goes away and the action buttons below it become:
- "Reset"
  - Clears the board.
- "Export"
  - Saves the pattern to a CSV file in the (project root)/tmp folder.
  - Exports are versioned by the current timestamp.

There isn't yet a way to consume the exported pattern, but that's
coming.

Also:
- Fix a bug where clicking and dragging from one cell to another on an
  active Board would trigger a click against the <tr> element. We've
  seen this before, but I didn't realize it could happen on the
  ToggleFlag action, too. (Perhaps it only can now that I've switched
  it to a left-click trigger in Draw Mode.)