Fix: Keep placed flags at Game-end, even on losing games
Previously, we were replacing flags with mines at the end of the game,
if the Game was a loss. This was just an oversight.

Also, dim the numbers text down a bit on Game end. This de-emphasizes
numbers, and adds emphasis to the flags/mines. (Which is good for
patterns.)