Allow import of mines from pattern CSV (dev-only)
Given a pattern CSV file (in tmp/patterns/), devs can now use the Rails
console to import the pattern as mines to a Standing-By Board. With the
caveat that, then, attempting to start the Game will fail without
commenting out the following line in Game#start:

```
Board::RandomlyPlaceMines.(board:, seed_cell:)
```

Will make this better in the future, once we determine a good way to
integrate patterns into normal game-play.