Add purpose-built Patterns, created via UI Portal
I was struggling with the overhead of mixing Draw Mode into the main
Game / Board. It was a dev-only feature, and what better way to make
a purpose-built version of the same thing than in the UI Portal (which
is already dev-only).

This approach does result in a lot of rebuilding of the same things that
already exist in the main Game / Board... but it also trims things down
a lot. So I'm happier this way.

Of note: The new "Board" stand-in is called Pattern. And a Pattern is,
in fact, an ActiveRecord model, backed by a database. But it doesn't
store off any Cells (or Cell equivalents) into the database, even for
rendering the Grid. Instead, we just store an array of Coordinates
representing flagged (virtual) Cells. This is more efficient,
storage-wise, no doubt. Which makes Patterns light-weight ... but also
easy to "Import" into a Board later. Because, Board "Import" will really
just be placing mines at the given Coordinates, which
Pattern#coordiantes_array natively represents. No muss, no fuss!