10% chance for "Random" to use a Pattern on Game start
So long as there are any Patterns to choose from, now:
- 10% of the time, starting a game with "Random" will randomly select a
  Pattern to apply to it
- This is effectively like an easter egg... though there's no way to
  "find" it, other than being dedicated to using the "Random" button.

Patterns are applied when the Board is created. So there is no safe
first-click. We could "fix" this by making an exception for first clicks
that would change the cell to not be a mine if it is one.