Show "Medal" icon as top-right nav user link in mobile
Previously, we just hid the User's name on mobile, because it didn't
fit along side our static nav links. But we do want to provide the
feature in mobile... so this is a "compromise" where we at least show
a clickable icon that navigates to the User's profile. And it looks
good, so we show the icon in web as well.