Don't transpose Patterns
To support this, it suddenly became obvious that handling the logic for
determining if the Grid should be transposed shouldn't actually belong
in Grid itself. As this is more of a view content / Board settings
concern. So, a better place to do this is in a new Games::GridContext
view model.

Games::GridContext is good at answering view context-related things,
like:
- allow_scrolling? (whether or not to add horizontal scroll bars to the
  grid)
- transpose? (whether or not to transpose the Grid)

This cleans up the Grid code quite a bit and feels like a solid win. i
wasn't sure how we would have made Grid aware of the current
Pattern::Settings otherwise...