Fix min-height of new / Custom Game forms
Previously, the transition from the New Game buttons to the New Custom
Game form would make the rest of the page jump down a bit within the
Game Show page (at the end of a Game). We mean to control this by
setting a min heights on the container element, though this wasn't
uniformly applied before.