Move "User Edit" under the Games resource layer
Users can only edit their username within the context of a Game. So the
resources/files/classes in the system should be similarly organized to
reflect this.

This effort started that simply... but I ended up adding a bunch of
one-off View Models to support the views (which I hadn't created back
when I should have) so that we didn't have to keep sending `game` and
`user` to every partial, ever. Instead, we can just cleanly send a
View Model that already knows about the Game and the User. Then, to make
this clean, had to add methods to the existing View Models to service
this... and on we went.
- A number of partials also moved from /games to /games/users/, but
  Git isn't tracking those movements well, so they show as individual
  deletes and re-creations.