Show Engagement Date in Game Results view
This is information we had from the Sweep Ops Archives listings. Just
reproducing it now in the Game Results view. This is useful regardless,
but especially now so that we can browse through past games via the
page sub-nav arrows (without having to go back to the index page first).