Improve jitteriness of Status Mojis "fleet"
Instead of showing a long list of ship emojis (which I still kind of
like better), we now show `⛴️ x1` or `⛴️ x2`, ... and so on to
represent how many players are participating (or watching) in the
War Room.

This is both a bit more intuitive and helps (but doesn't yet solve) the
jitters we get when players unsubscribe/resubscribe from/to the
`:current_game` Turbo Stream channel. It was worse before this change
because each new ship emoji would add to the list and move the entire
status line over. Vs now it just updates the number e.g. from
2 -> 1 -> 2 real quick.
- Unfortunately, this happens on every game board click... Apparently
  the page refresh and/or the AJAX request causes an
  un-subscription/re-subscription event in the Turbo Stream... which
  we dutifully report on via Games::Show#broadcast_players_count_update.
- I suppose things could be jittery again if/when we reach the border
  between single and double-player/participant counts.

Also, I've changed the source of the "player count" value that is
broadcast in subscription/unsubscription updates. We now pull this from
DutyRoster.count instead of using Array#size based on the the result of
the add/remove command. This probably won't change anything, but it
feels like it could leave a bit of space for DutyRoster to do something
to alleviate this issue somehow (in the future?).